Made endpoints in order to see the details page of a specific movie. This detailspage is needed in order to implement the watchlist feature

+++ app/src/main/java/com/example/riseandroid/LumiereApplication.kt
@@ -9,8 +9,6 @@ import android.content.Context
 import android.content.Intent
 import com.example.riseandroid.notifications.NotificationReceiver
 import java.util.Calendar
-import android.app.NotificationManager
-import android.os.Build
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 
@@ -23,7 +21,7 @@ class LumiereApplication : Application() {
         container = DefaultAppContainer(context = applicationContext)
     }
 
-    fun scheduleNotification(context: Context, movieId: Long, movieName: String, location: String, date: String, dateTime: Calendar) {
+    fun scheduleNotification(context: Context, movieId: Int, movieName: String, location: String, date: String, dateTime: Calendar) {
         val intent = Intent(context, NotificationReceiver::class.java).apply {
             putExtra("MOVIE_ID", movieId)
             putExtra("MOVIE_NAME", movieName)
@@ -43,7 +41,7 @@ class LumiereApplication : Application() {
 
 
 
-    fun displayImmediateNotification(context: Context, movieId: Long, movieName: String, location: String, date: String) {
+    fun displayImmediateNotification(context: Context, movieId: Int, movieName: String, location: String, date: String) {
         if (NotificationManagerCompat.from(context).areNotificationsEnabled()) {
             val notification = NotificationCompat.Builder(context, "REMINDER_CHANNEL")
                 .setSmallIcon(R.drawable.ic_launcher_foreground) // Replace with your app's icon
